+++ Demo/Application/Features/Views/CardFormView.swift
@@ -1,46 +1,129 @@
 import UIKit
 
-@objc public class CardFormView: UIView {
-
-    // MARK: 
+class CardFormView: UIView {
 
     // MARK: - Public Properties
 
+    @objc var cardNumber: String? {
+        return cardNumberTextField.text != "" ? cardNumberTextField.text : nil
+    }
+    @objc var expirationMonth: String? {
+        return parseExpirationDate()?.month
+    }
+    @objc var expirationYear: String? {
+        return parseExpirationDate()?.year
+    }
+    @objc var cvv: String? {
+        return cvvTextField.text != "" ? cvvTextField.text : nil
+    }
+    @objc var postalCode: String? {
+        return postalCodeTextField.text != "" ? postalCodeTextField.text : nil
+    }
+    @objc var phoneNumber: String? {
+        return phoneNumberTextField.text != "" ? phoneNumberTextField.text : nil
+    }
+
+    // MARK: - Public stuff
+
     @objc var hidePhoneNumberField: Bool = false {
         didSet {
-            phoneNumberLabel.isHidden = self.hidePhoneNumberField
             phoneNumberTextField.isHidden = self.hidePhoneNumberField
         }
     }
-
     @objc var hidePostalCodeField: Bool = false {
         didSet {
-            postalCodeLabel.isHidden = self.hidePostalCodeField
             postalCodeTextField.isHidden = self.hidePostalCodeField
         }
     }
-
-    @objc var hideCVVTextField: Bool = false {
+    @objc var hideCVVField: Bool = false {
         didSet {
-            cvvLabel.isHidden = self.hideCVVTextField
-            cvvTextField.isHidden = self.hideCVVTextField
+            cvvTextField.isHidden = self.hideCVVField
         }
     }
 
-    // MARK: - IBOutlets
+    // MARK: - Internal UI Properties
 
-    @IBOutlet weak var cardNumberTextField: UITextField!
-    @IBOutlet weak var expirationMonthTextField: UITextField!
-    @IBOutlet weak var expirationYearTextField: UITextField!
-    @IBOutlet weak var cvvTextField: UITextField!
-    @IBOutlet weak var postalCodeTextField: UITextField!
-    @IBOutlet weak var phoneNumberTextField: UITextField!
+    @objc var cardNumberTextField: UITextField = UITextField() // exposed for UnionPay demo
+    var expirationTextField: UITextField = UITextField()
+    var cvvTextField: UITextField = UITextField()
+    var postalCodeTextField: UITextField = UITextField()
+    var phoneNumberTextField: UITextField = UITextField()
+    var stackView: UIStackView = UIStackView()
 
-    @IBOutlet weak var cardNumberLabel: UILabel!
-    @IBOutlet weak var expirationMonthLabel: UILabel!
-    @IBOutlet weak var expirationYearLabel: UILabel!
-    @IBOutlet weak var cvvLabel: UILabel!
-    @IBOutlet weak var postalCodeLabel: UILabel!
-    @IBOutlet weak var phoneNumberLabel: UILabel!
+    let padding: CGFloat = 10
+
+    // MARK: Initializers
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+        configureStackView()
+        configureStackViewComponents()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    // MARK: - Layout UI
+
+    func configureStackView() {
+        stackView.axis = .vertical
+        stackView.distribution = .fillEqually
+        stackView.alignment = .leading
+        stackView.spacing = 10
+
+        stackView.translatesAutoresizingMaskIntoConstraints = false
+        addSubview(stackView)
+
+        NSLayoutConstraint.activate([
+            stackView.topAnchor.constraint(equalTo: safeAreaLayoutGuide.topAnchor, constant: padding),
+            stackView.bottomAnchor.constraint(equalTo: safeAreaLayoutGuide.bottomAnchor, constant: -padding),
+            stackView.leadingAnchor.constraint(equalTo: safeAreaLayoutGuide.leadingAnchor, constant: padding),
+            stackView.trailingAnchor.constraint(equalTo: safeAreaLayoutGuide.trailingAnchor, constant: -padding),
+        ])
+    }
+
+    func configureStackViewComponents() {
+        cardNumberTextField.placeholder = "Card Number"
+        expirationTextField.placeholder = "MM/YY"
+        cvvTextField.placeholder = "CVV"
+        postalCodeTextField.placeholder = "Postal Code"
+        phoneNumberTextField.placeholder = "Phone Number"
+
+        cardNumberTextField.translatesAutoresizingMaskIntoConstraints = false
+        expirationTextField.translatesAutoresizingMaskIntoConstraints = false
+        cvvTextField.translatesAutoresizingMaskIntoConstraints = false
+        postalCodeTextField.translatesAutoresizingMaskIntoConstraints = false
+        phoneNumberTextField.translatesAutoresizingMaskIntoConstraints = false
+
+        stackView.addArrangedSubview(cardNumberTextField)
+        stackView.addArrangedSubview(expirationTextField)
+        stackView.addArrangedSubview(cvvTextField)
+        stackView.addArrangedSubview(postalCodeTextField)
+        stackView.addArrangedSubview(phoneNumberTextField)
+
+        NSLayoutConstraint.activate([
+            cardNumberTextField.widthAnchor.constraint(equalTo: stackView.widthAnchor),
+            expirationTextField.widthAnchor.constraint(equalTo: stackView.widthAnchor),
+            cvvTextField.widthAnchor.constraint(equalTo: stackView.widthAnchor),
+            postalCodeTextField.widthAnchor.constraint(equalTo: stackView.widthAnchor),
+            phoneNumberTextField.widthAnchor.constraint(equalTo: stackView.widthAnchor)
+        ])
+    }
+
+    // MARK: - Helpers
+
+    func parseExpirationDate() -> (month: String, year: String)? {
+        guard let expirationDate = expirationTextField.text, expirationTextField.text != "" else {
+            return nil
+        }
+
+        let dateComponents = expirationDate.components(separatedBy: "/")
+        if dateComponents.count != 2 {
+            return nil
+        } else {
+            return (dateComponents.first!, dateComponents.last!)
+        }
+    }
 
 }
